Transform data using ArcSinh transformation

Function to transform data in selected columns using ArcSinh transformation with a specified co-factor.

Usage

do.asinh(dat, use.cols, cofactor=5, append.cf=FALSE,
reduce.noise=FALSE)

Arguments

dat

NO DEFAULT. data.table object. Input data.

use.cols

NO DEFAULT. Vector of character column names. The columns to transform. These columns will be transformed and added to the data.table as new columns.

cofactor

DEFAULT = 5. Co-factor to use for arcsinh transformation. Can be list of co-factors that matches the columns in use.cols. If specifying a list of co-factors make sure each column in use.cols is assigned a co-factor (repeat values if necessary).

append.cf

DEFAULT = FALSE. Appends the co-factor used to the end of the name of the transformed columns.

reduce.noise

DEFAULT = FALSE. This is an experimental calculation which should reduce noise from negative values. Use with caution.

Value

A data.table with new columns added, that contain asinh transformed data.