do.clip - Clips data using a specified lower and upper value

This function allows you to clip data using specified lower and upper values. E.g. if the upper value is set to 1000, then any values above 1000 will be converted to 1000. No rows are lost, rather the values in those rows are converted.

Usage

do.clip(dat, use.cols, min.value = NULL, max.value = NULL, append.name = '_clipped')

Arguments

dat

NO DEFAULT. data.table Input sample.

use.cols

NO DEFAULT. Columns to clip

min.value

DEFAULT = NULL. Specify the value for which values below will be converted to. If set to NULL, no clipping will occur.

max.value

DEFAULT = NULL. Specify the value for which values above will be converted to. If set to NULL, no clipping will occur.

append.name

DEFAULT = '_clipped'. Text appended to new column names after clipping.

Value

A data.table with new columns added, that contain the clipped data.

References

https://sydneycytometry.org.au/spectre

See also