do.zscore - Calculate z.score for a dataset

do.zscore - Calculate z.score for a dataset

Usage

do.zscore(dat, use.cols)

Arguments

dat

NO DEFAULT. A data.table (or data.frame) containing the data to be converted to z-scores. Z-score transformed values will be added as new columns.

use.cols

NO DEFAULT. The columns to be used for z-score calculations.

append.name

DEFAULT = '_zscore'. Text to be appended to the end of the new z-score transformed columns.

replace

DEFAULT = FALSE. If FALSE, appends new columns to the data.table. If TRUE, replaces the values in the existing columns with the z-score tranformed values, and does not change the column names.

Value

Returns a new data.table with z-score calculations for each selected column

Examples

do.zscore(dat = Spectre::demo.clustered, use.cols = c("NK11", "CD4"))
#>                  FileName      NK11        CD3     CD45       Ly6G    CD11b
#>                    <char>     <num>      <num>    <num>      <num>    <num>
#>      1:   CNS_Mock_01.csv   42.3719  40.098700  6885.08  -344.7830 14787.30
#>      2:   CNS_Mock_01.csv   42.9586 119.014000  1780.29  -429.6650  5665.73
#>      3:   CNS_Mock_01.csv   59.2366 206.238000 10248.30 -1603.8400 19894.30
#>      4:   CNS_Mock_01.csv  364.9480  -0.233878  3740.04  -815.9800  9509.43
#>      5:   CNS_Mock_01.csv  440.2470  40.035200  9191.38    40.5055  5745.82
#>     ---                                                                    
#> 169000: CNS_WNV_D7_06.csv  910.8890  72.856100 31466.20  -316.5570 28467.80
#> 169001: CNS_WNV_D7_06.csv  -10.2642  64.188700 45188.00  -540.5140 22734.00
#> 169002: CNS_WNV_D7_06.csv -184.2910  -9.445650 11842.60   -97.9383 17237.00
#> 169003: CNS_WNV_D7_06.csv  248.3860 229.986000 32288.20  -681.1630 19255.80
#> 169004: CNS_WNV_D7_06.csv  738.9810  95.470300 46185.10 -1004.6000 22957.80
#>              B220      CD8a       Ly6C       CD4  NK11_asinh    CD3_asinh
#>             <num>     <num>      <num>     <num>       <num>        <num>
#>      1:  -40.2399   83.7175   958.7000  711.0720  0.04235923  0.040087962
#>      2:   86.6673   34.7219   448.2590  307.2720  0.04294540  0.118734817
#>      3:  427.8310  285.8800  1008.8300  707.0940  0.05920201  0.204803270
#>      4:  182.4200  333.6050   440.0710  249.7840  0.35729716 -0.000233878
#>      5: -211.6940  149.2200    87.4815  867.5700  0.42713953  0.040024513
#>     ---                                                                  
#> 169000:   -7.7972 -271.8040 12023.7000 1103.0500  0.81693878  0.072791800
#> 169001:  202.4110 -936.4920  4188.3300  315.9400 -0.01026402  0.064144703
#> 169002:  123.4760 -219.9320  8923.4000 -453.4640 -0.18326344 -0.009445510
#> 169003: -656.0540 -201.5880 10365.7000   61.6765  0.24590035  0.228005328
#> 169004: -661.6280   72.3356  9704.4700  -31.8532  0.68430866  0.095325863
#>         CD45_asinh  Ly6G_asinh CD11b_asinh   B220_asinh  CD8a_asinh Ly6C_asinh
#>              <num>       <num>       <num>        <num>       <num>      <num>
#>      1:   2.627736 -0.33829345    3.388057 -0.040229048  0.08362002  0.8518665
#>      2:   1.340828 -0.41743573    2.435282  0.086559169  0.03471493  0.4344615
#>      3:   3.022631 -1.25101677    3.684212  0.415750122  0.28212257  0.8876036
#>      4:   2.029655 -0.74509796    2.948184  0.181423123  0.32770787  0.4269784
#>      5:   2.914359  0.04049443    2.449108 -0.210143906  0.14867171  0.0873703
#>     ---                                                                       
#> 169000:   4.142314 -0.31149515    4.042229 -0.007797121 -0.26856390  3.1817517
#> 169001:   4.504101 -0.51715205    3.817492  0.201053740 -0.83574631  2.1394053
#> 169002:   3.166628 -0.09778240    3.541046  0.123164374 -0.21819650  2.8849492
#> 169003:   4.168089 -0.63716643    3.651633 -0.616293228 -0.20024703  3.0339681
#> 169004:   4.525922 -0.88462254    3.827279 -0.620947819  0.07227267  2.9683779
#>           CD4_asinh     Sample  Group  Batch FlowSOM_cluster
#>               <num>     <char> <char> <char>           <num>
#>      1:  0.66171351 01_Mock_01   Mock      A              23
#>      2:  0.30263135 01_Mock_01   Mock      A              55
#>      3:  0.65846851 01_Mock_01   Mock      A              64
#>      4:  0.24725691 01_Mock_01   Mock      A              53
#>      5:  0.78456678 01_Mock_01   Mock      A             110
#>     ---                                                     
#> 169000:  0.95239703  12_WNV_06    WNV      A              72
#> 169001:  0.31090687  12_WNV_06    WNV      A              46
#> 169002: -0.43920651  12_WNV_06    WNV      A             133
#> 169003:  0.06163746  12_WNV_06    WNV      A             133
#> 169004: -0.03184782  12_WNV_06    WNV      A             103
#>         FlowSOM_metacluster        Population     UMAP_X    UMAP_Y NK11_zscore
#>                      <fctr>            <char>      <num>     <num>       <num>
#>      1:                   2         Microglia -2.3603757  6.201213 -0.31824725
#>      2:                   2         Microglia  2.7505242  7.119595 -0.31760502
#>      3:                   2         Microglia -2.9486033  4.012670 -0.29978628
#>      4:                   2         Microglia  0.6482904  6.481466  0.03486128
#>      5:                   4          NK cells -2.3941295  6.975885  0.11728748
#>     ---                                                                       
#> 169000:                   3 Infil Macrophages -2.9640724 -5.058265  0.63247663
#> 169001:                   3 Infil Macrophages -1.2644785 -3.555824 -0.37586546
#> 169002:                   3 Infil Macrophages -2.3592682 -2.429467 -0.56636423
#> 169003:                   3 Infil Macrophages -1.9531062 -4.049705 -0.09273353
#> 169004:                   3 Infil Macrophages -0.7404098 -4.686928  0.44429721
#>          CD4_zscore
#>               <num>
#>      1: -0.08665135
#>      2: -0.14668865
#>      3: -0.08724280
#>      4: -0.15523601
#>      5: -0.06338311
#>     ---            
#> 169000: -0.02837176
#> 169001: -0.14539989
#> 169002: -0.25979547
#> 169003: -0.18320398
#> 169004: -0.19711005