package.check - a function to check the installation of all required packages.

This function allows you to check to see if all the common use packages dependencies for Spectre are installed. See https://sydneycytometry.org.au/spectre for usage instructions and vignettes.

Usage

package.check(type = "general")

Arguments

type

DEFAULT = "general". If "general", then checks for the packages required for general Spectre usage. If "spatial", then checks for additional packages required for spatial analysis. If "ML", then checks for additional packages required for machine-learing functionality.

Value

returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.

References

https://sydneycytometry.org.au/spectre

Author

Thomas M Ashhurst, thomas.ashhurst@sydney.edu.au

Examples

package.check()
#> Package: Spectre
#>  -- Version:           1.3.0
#>  -- Install date:      2025-09-18 12:28:20 UTC; runner
#>  -- Install source:    local
#>  -- R version:         R version 4.5.1 (2025-06-13)
#>  -- OS:                Ubuntu 24.04.3 LTS
#>  -- OS detail:         x86_64-pc-linux-gnu
#>  -- Library path(s):      
#>         /home/runner/work/_temp/Library
#>         /opt/R/4.5.1/lib/R/site-library
#>         /opt/R/4.5.1/lib/R/library
#>                
#> Checking dependency packages...
#>  -- All required packages are installed.