package.install - a function to install packages required for Spectre.

This function allows you to check to see if all the common use packages dependencies for Spectre are installed. Will only install if the package has not been installed, will not update packages. See https://sydneycytometry.org.au/spectre for usage instructions and vignettes.

Usage

package.install(type = "general", update = FALSE)

Arguments

type

DEFAULT = "general". If "general", then checks for the packages required for general Spectre usage. If "spatial", then checks for additional packages required for spatial analysis.

update

DEFAULT = FALSE. If FALSE, will only install packages that are not already installed – no updates of packages will be performed. If TRUE, will install and update packages.

Value

returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.

References

https://sydneycytometry.org.au/spectre

Author

Thomas M Ashhurst, thomas.ashhurst@sydney.edu.au

Examples

package.install()
#> Loading required package: plyr
#> 
#> Attaching package: ‘plyr’
#> The following object is masked from ‘package:ggpubr’:
#> 
#>     mutate
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:plyr’:
#> 
#>     arrange, count, desc, failwith, id, mutate, rename, summarise,
#>     summarize
#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine
#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, setequal, union
#> The following object is masked from ‘package:generics’:
#> 
#>     explain
#> The following object is masked from ‘package:flowCore’:
#> 
#>     filter
#> The following objects are masked from ‘package:data.table’:
#> 
#>     between, first, last
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#> Loading required package: tidyr
#> Loading required package: rstudioapi
#> Loading required package: Rtsne
#> Loading required package: umap
#> Loading required package: gridExtra
#> 
#> Attaching package: ‘gridExtra’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     combine
#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine
#> The following object is masked from ‘package:BiocGenerics’:
#> 
#>     combine
#> Loading required package: factoextra
#> Welcome! Want to learn more? See two factoextra-related books at https://goo.gl/ve3WBa
#> Loading required package: reticulate
#> Loading required package: caret
#> Loading required package: lattice
#> 
#> Attaching package: ‘caret’
#> The following object is masked from ‘package:generics’:
#> 
#>     train
#> Loading required package: class
#> Loading required package: flowViz
#> Loading required package: FlowSOM
#> Loading required package: igraph
#> 
#> Attaching package: ‘igraph’
#> The following object is masked from ‘package:class’:
#> 
#>     knn
#> The following object is masked from ‘package:tidyr’:
#> 
#>     crossing
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     as_data_frame, groups, union
#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     normalize, path, union
#> The following objects are masked from ‘package:generics’:
#> 
#>     components, union
#> The following object is masked from ‘package:flowCore’:
#> 
#>     normalize
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
#> Thanks for using FlowSOM. From version 2.1.4 on, the scale 
#> parameter in the FlowSOM function defaults to FALSE
#> 
#> Attaching package: ‘FlowSOM’
#> The following object is masked from ‘package:class’:
#> 
#>     SOM
package.install(type = 'spatial')
#> Loading required package: raster
#> Loading required package: sp
#> 
#> Attaching package: ‘raster’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
#> Loading required package: tiff
#> Loading required package: rgeos
#> Warning: there is no package called ‘rgeos’
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Warning: package ‘rgeos’ is not available for this version of R
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Loading required package: exactextractr
#> Loading required package: sf
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
#> Loading required package: stars
#> Loading required package: abind
#> Loading required package: qs
#> qs 0.27.3. Announcement: https://github.com/qsbase/qs/issues/103
#> Loading required package: s2