run.align - Run the alignment model on a target data.table

This function allows you to prepare reference data ahead of performing batch alignment, using on of the CytofBatchAdjust functions

Usage

run.align()

Arguments

ref.dat

NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm

target.dat

NO DEFAULT.A data.table consisting of the 'target' data you will use to align the data

batch.col

NO DEFAULT. Column name of the data.table that contains batch labels

align.cols

NO DEFAULT. Vector of column names to align.

method

DEFAULT = 'quantile'. Can be 'quantile', 'SD', or a percentile indicated as '50p' (50th percentile) or '95p' (95th percentile) etc.

append.name

DEFAULT = "_aligned". This will be appended to the column names containing the aligned data

dir

DEFAULT = getwd(). Sets the working directory to operate from. Because this function involves some reading/writing of files, it's best to set this to somewhere static in case the active working directory moves to a subfolder, and then doesn't return because the function runs into an error.

mem.ctrl

DEFAULT = TRUE. Allows the function to clear held memory on occasion.

Value

Returns a data.table with aligned data added in new columns.

References

Ashhurst, T. M., et al. (2019). https://www.ncbi.nlm.nih.gov/pubmed/31077106

Author

Thomas M Ashhurst, thomas.ashhurst@sydney.edu.au