train.cytonorm - Prepare reference data into a FlowSOM object

This function allows you to learn the conversions to use for batch alignment.

Usage

train.align()

Arguments

method

DEFAULT = 'cytonorm'. In future additional methods for batch alignment will be added.

cytonorm.goal

DEFAULT = 'mean'. Target values for alignment. Can be 'mean' for the average of all batches, or a specific batch can be selected.

cytonorm.nQ

DEFAULT = 101. Number of quantiles.

dir

DEFAULT = getwd(). Sets the working directory to operate from. Because this function involves some reading/writing of files, it's best to set this to somewhere static in case the active working directory moves to a subfolder, and then doesn't return because the function runs into an error.

mem.ctrl

DEFAULT = TRUE. Allows the function to clear held memory on occasion.

dat

NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm

cellular.cols

NO DEFAULT. A vector of column names from the data.table that contain the markers to be aligned

Value

Returns the alignment model object, with an additional 'conversions' element added.

References

Ashhurst, T. M., et al. (2019). https://www.ncbi.nlm.nih.gov/pubmed/31077106

Author

Thomas M Ashhurst, thomas.ashhurst@sydney.edu.au